Published library to Center public repo and made required project changes in build files.
