+++ UniversalAdapter/.gitignore
