--- UniversalAdapter/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=UniversalAdapter
-POM_PACKAGING=aar
-POM_ARTIFACT_ID=UniversalAdapter
