--- UniversalAdapter/src/androidTest/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.raizlabs.universaladapter"
-          xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-      />
-
-</manifest>
