+++ UniversalAdapter/src/androidTest/java/com/raizlabs/universaladapter/test/VanillaAdapterTest.java
@@ -3,12 +3,15 @@ package com.raizlabs.universaladapter.test;
 import android.view.View;
 import android.widget.LinearLayout;
 
-import com.raizlabs.android.coreutils.util.observable.lists.ListObserver;
-import com.raizlabs.android.coreutils.util.observable.lists.ListObserverListener;
+import com.raizlabs.coreutils.util.observable.lists.ListObserver;
+import com.raizlabs.coreutils.util.observable.lists.ListObserverListener;
 import com.raizlabs.universaladapter.ViewHolder;
 import com.raizlabs.universaladapter.converter.UniversalAdapterTestCase;
 
-import static com.raizlabs.universaladapter.test.Constants.*;
+import static com.raizlabs.universaladapter.test.Constants.INDEX_CHANGED;
+import static com.raizlabs.universaladapter.test.Constants.INDEX_GENERIC;
+import static com.raizlabs.universaladapter.test.Constants.INDEX_INSERTED;
+import static com.raizlabs.universaladapter.test.Constants.INDEX_REMOVED;
 
 /**
  * Description: Tests to ensure vanilla adapters (not merged) handle item types and other data appropriately.
