+++ UniversalAdapter/src/main/AndroidManifest.xml
@@ -1,12 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.raizlabs.universaladapter"
-    android:versionCode="1"
-    android:versionName="1.0" >
+          package="com.raizlabs.universaladapter">
 
-    <uses-sdk
-        android:minSdkVersion="8"
-        android:targetSdkVersion="21" />
+    <application android:allowBackup="true"
+                 android:label="@string/app_name">
 
-    <application />
+    </application>
 
 </manifest>
