+++ app/build.gradle
@@ -1,14 +1,15 @@
-apply plugin: 'com.android.library'
+apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion 22
     buildToolsVersion "22.0.1"
 
     defaultConfig {
+        applicationId "com.raizlabs.universaladapter.app"
         minSdkVersion 15
         targetSdkVersion 22
         versionCode 1
-        versionName VERSION_NAME
+        versionName "1.0"
     }
     buildTypes {
         release {
@@ -20,9 +21,8 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:22.2.0'
+    compile 'com.android.support:appcompat-v7:22.+'
     compile 'com.android.support:recyclerview-v7:22.2.0'
     compile 'com.raizlabs:CoreUtils:1.1.0'
+    compile project(':UniversalAdapter')
 }
-
-apply from: 'https://raw.githubusercontent.com/Raizlabs/maven-releases/master/raizlabs_bintray_upload.gradle'
