+++ app/src/main/java/com/raizlabs/universaladapter/app/AdapterActivity.java
@@ -1,9 +1,9 @@
-package com.raizlabs.universaladapter.sample.multiple;
+package com.raizlabs.universaladapter.app;
 
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import android.support.v4.app.FragmentActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -20,24 +20,18 @@ import com.raizlabs.universaladapter.converter.ItemLongClickedListener;
 import com.raizlabs.universaladapter.converter.UniversalAdapter;
 import com.raizlabs.universaladapter.converter.UniversalConverter;
 import com.raizlabs.universaladapter.converter.UniversalConverterFactory;
-import com.raizlabs.universaladapter.sample.AdapterActivity;
-import com.raizlabs.universaladapter.sample.R;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
-/**
- * Description: Provides multiple kinds of adapters with headers and footers of all kinds.
- */
-public class MultipleItemTypesHFActivity extends AppCompatActivity {
+public class AdapterActivity extends FragmentActivity {
 
     static final String INTENT_LAYOUT_RESOURCE = "LayoutResource";
 
     static final String INTENT_USE_HF = "UseHF";
 
     public static Intent getLaunchIntent(Context context, int layoutResId) {
-        Intent intent = new Intent(context, MultipleItemTypesHFActivity.class);
+        Intent intent = new Intent(context, AdapterActivity.class);
         intent.putExtra(INTENT_LAYOUT_RESOURCE, layoutResId);
         return intent;
     }
@@ -48,7 +42,7 @@ public class MultipleItemTypesHFActivity extends AppCompatActivity {
         return intent;
     }
 
-    private UniversalConverter<Object, ViewHolder> converter;
+    private UniversalConverter<Object, UniversalHolder> converter;
 
     private ListBasedAdapter<Object, ? extends ViewHolder> adapter;
 
@@ -58,57 +52,80 @@ public class MultipleItemTypesHFActivity extends AppCompatActivity {
         int layout = getIntent().getIntExtra(INTENT_LAYOUT_RESOURCE, R.layout.activity_viewpager);
         setContentView(layout);
 
-        adapter = new MultipleItemTypesAdapter();
-
-        List<Object> itemsList = new ArrayList<>();
-        Random random = new Random();
-        for(int i = 0; i < 50; i++) {
-            int type = random.nextInt(2);
-            if(type == MultipleItemTypesAdapter.VIEW_TYPE_ICON) {
-                itemsList.add(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
-            } else {
-                itemsList.add("" + i);
+        adapter = new ListBasedAdapter<Object, UniversalHolder>() {
+            @Override
+            protected void onBindViewHolder(UniversalHolder viewHolder, Object s, int position) {
+                viewHolder.Gibberish.setText(s.toString());
             }
-        }
+
+            @Override
+            protected UniversalHolder onCreateViewHolder(ViewGroup parent, int itemType) {
+                return new UniversalHolder(inflateView(parent, R.layout.list_item_adapter));
+            }
+
+            @Override
+            protected void onBindHeaderViewHolder(ViewHolder holder, int position) {
+                HeaderHolder headerHolder = (HeaderHolder) holder;
+                headerHolder.Image.setImageResource(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
+            }
+
+            @Override
+            protected void onBindFooterViewHolder(ViewHolder holder, int position) {
+                FooterHolder footerHolder = (FooterHolder) holder;
+                footerHolder.Dummy.setText("Footer: " + position);
+            }
+        };
 
         if (getIntent().hasExtra(INTENT_USE_HF)) {
             for (int i = 0; i < 5; i++) {
                 HeaderHolder headerHolder = new HeaderHolder(LayoutInflater.from(this).inflate(R.layout.list_item_image, null));
-                headerHolder.Image.setImageResource(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
                 adapter.addHeaderHolder(headerHolder);
             }
 
             for (int i = 0; i < 5; i++) {
                 FooterHolder footerHolder = new FooterHolder(LayoutInflater.from(this).inflate(R.layout.list_item_text, null));
-                footerHolder.Dummy.setText(i + "");
                 adapter.addFooterHolder(footerHolder);
             }
         }
 
-        adapter.loadItemList(itemsList);
+        List<Object> dummyTitles = new ArrayList<>();
+        for (int i = 0; i < 50; i++) {
+            dummyTitles.add(String.valueOf(i));
+        }
+
+        adapter.loadItemList(dummyTitles);
 
         ViewGroup viewGroup = (ViewGroup) findViewById(R.id.content);
         if (viewGroup instanceof RecyclerView) {
             ((RecyclerView) viewGroup).setLayoutManager(new LinearLayoutManager(viewGroup.getContext()));
         }
 
-        converter = (UniversalConverter<Object, ViewHolder>) UniversalConverterFactory.createGeneric(adapter,
-                                                                                                          viewGroup);
-        converter.setItemClickedListener(new ItemClickedListener<Object, ViewHolder>() {
+        converter = (UniversalConverter<Object, UniversalHolder>) UniversalConverterFactory.createGeneric(adapter, viewGroup);
+        converter.setItemClickedListener(new ItemClickedListener<Object, UniversalHolder>() {
             @Override
-            public void onItemClicked(UniversalAdapter<Object, ViewHolder> adapter, Object o,
-                                      ViewHolder holder, int position) {
+            public void onItemClicked(UniversalAdapter<Object, UniversalHolder> adapter, Object o,
+                                      UniversalHolder holder, int position) {
                 Toast.makeText(holder.itemView.getContext(), "Clicked on :" + position, Toast.LENGTH_SHORT).show();
             }
         });
-        converter.setItemLongClickedListener(new ItemLongClickedListener<Object, ViewHolder>() {
+        converter.setItemLongClickedListener(new ItemLongClickedListener<Object, UniversalHolder>() {
             @Override
-            public boolean onItemLongClicked(UniversalAdapter<Object, ViewHolder> adapter, Object o,
-                                             ViewHolder holder, int position) {
+            public boolean onItemLongClicked(UniversalAdapter<Object, UniversalHolder> adapter, Object o,
+                                             UniversalHolder holder, int position) {
                 Toast.makeText(holder.itemView.getContext(), "Long clicked on: " + position, Toast.LENGTH_SHORT).show();
                 return false;
             }
         });
+    }
+
+    private static class UniversalHolder extends ViewHolder {
+
+        TextView Gibberish;
+
+        public UniversalHolder(View itemView) {
+            super(itemView);
+            Gibberish = ((TextView) itemView);
+        }
     }
 
     private static class HeaderHolder extends ViewHolder {
