+++ app/src/main/res/layout/activity_linearlayout.xml
