+++ build.gradle
@@ -1,23 +1,21 @@
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+
 buildscript {
     repositories {
         jcenter()
-        maven { url "https://raw.github.com/Raizlabs/maven-releases/master/releases" }
-
     }
     dependencies {
-        classpath 'com.raizlabs:Griddle:1.0.3'
         classpath 'com.android.tools.build:gradle:1.2.3'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.2'
         classpath 'com.github.dcendents:android-maven-plugin:1.2'
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
     }
 }
 
 allprojects {
-    group = GROUP
-    version = VERSION
-
     repositories {
         jcenter()
-        maven { url "https://raw.github.com/Raizlabs/maven-releases/master/releases" }
     }
 }
