+++ gradle.properties
@@ -1,17 +1,4 @@
-VERSION=2.2.0
-VERSION_CODE=1
-GROUP=com.raizlabs
-
-POM_NAME=UniversalAdapter
-POM_PACKAGING=aar
-POM_ARTIFACT_ID=UniversalAdapter
-POM_DESCRIPTION=UniversalAdapter
-POM_URL=https://github.com/Raizlabs/UniversalAdapter
-POM_SCM_URL=https://github.com/Raizlabs/UniversalAdapter.git
-POM_SCM_CONNECTION=scm:git@github.com:Raizlabs/UniversalAdapter.git
-POM_SCM_DEV_CONNECTION=scm:git@github.com:Raizlabs/UniversalAdapter.git
-POM_LICENCE_NAME=The MIT License
-POM_LICENCE_URL=http://opensource.org/licenses/MIT
-POM_LICENCE_DIST=repo
-
-deployPath = ../../maven-releases/releases
+VERSION_NAME=2.2.0
+ARTIFACT_NAME = UniversalAdapter
+GIT_URL = https://github.com/Raizlabs/UniversalAdapter.git
+SITE_URL = https://github.com/Raizlabs/UniversalAdapter
