--- sample/build.gradle
@@ -1,21 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'com.raizlabs.griddle'
-
-android {
-    compileSdkVersion 22
-    buildToolsVersion "22.0.1"
-
-    defaultConfig {
-        applicationId "com.raizlabs.universaladapter.sample"
-        minSdkVersion 14
-        targetSdkVersion 22
-        versionCode 1
-        versionName "1.0"
-    }
-}
-
-dependencies {
-    nsMod 'com.android.support:{appcompat-v7, recyclerview-v7}:22.1.0'
-    compile project(':UniversalAdapter')
-    mod 'com.raizlabs.android:{CoreUtils}:1.0.0'
-}
